test: decouple and migrate test suite into modular cases
